update types in line with changes made as part of the 1.0.9 deployment

+++ uk.ac.diamond.ispyb.api/src/uk/ac/diamond/ispyb/api/IspybPlateApi.java
@@ -27,11 +27,11 @@ public interface IspybPlateApi extends Closeable {
 
 	void insertContainerError(String barcode, String error, int severity, String stackTrace) throws SQLException;
 
-	public List<DataCollectionInfo> retrieveDataCollectionInfosForSubsample(int id);
+	public List<DataCollectionInfo> retrieveDataCollectionInfosForSubsample(long id);
 	
 	void clearContainerError(String barcode) throws SQLException;
 
-	public int upsertSampleImageAnalysis(SampleImageAnalysis sampleImageAnalysis);
+	public Long upsertSampleImageAnalysis(SampleImageAnalysis sampleImageAnalysis);
 	
 	boolean isConnected();
 }
